requirements.txt
# Requirements for Autonomous Research Agent

# Core dependencies
langgraph>=0.0.32
langchain>=0.1.0
langchain-google-genai>=0.0.1
langchain-core>=0.1.0

# LLM and Embeddings
google-generativeai>=0.3.0

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Async Support
aiohttp>=3.9.0
asyncio-contextmanager>=1.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
psycopg-binary>=3.0.0

# Vector Stores
pinecone-client>=3.0.0

# Web Search
tavily-python>=0.1.0
requests>=2.31.0

# Frontend
streamlit>=1.28.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utils
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
isort>=5.12.0

# Logging
python-json-logger>=2.0.7
